Environments improvements
