+++ assets/README.md
@@ -1,0 +1,1 @@
+Flutter application assets are taken from the assets folder. This folder contains the assets used by the application, such as images, fonts, and other files.
