+++ environments/README.md
@@ -1,0 +1,8 @@
+Application environments are taken from the environments folder. 
+This folder contains the environment variables used by the application, such as API keys, and other files.
+
+The default environment variables for all environments are defined in `default.env`.
+
+To specify environment variables for a specific environment, you can create a file with the name of the environment, for example, dev.env for development environment.
+To specify private environment variables for a specific environment, you can create a file with the name of the environment and the suffix .private, for example, dev.private.env for development environment.
+
