+++ lib/gen/assets.gen.dart
@@ -9,23 +9,8 @@
 
 import 'package:flutter/widgets.dart';
 
-class $AssetsEnvironmentsGen {
-  const $AssetsEnvironmentsGen();
-
-  /// File path: assets/environments/.env
-  String get env => 'assets/environments/.env';
-
-  /// File path: assets/environments/development.env
-  String get development => 'assets/environments/development.env';
-
-  /// List of all assets
-  List<String> get values => [env, development];
-}
-
 class Assets {
   Assets._();
-
-  static const $AssetsEnvironmentsGen environments = $AssetsEnvironmentsGen();
 }
 
 class AssetGenImage {
